requirements.txt
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.7
python-dateutil>=2.8.2
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.7b0
isort>=5.9.3
mypy>=0.910
typer>=0.9.0
fastapi>=0.103.0
uvicorn>=0.23.0
jinja2>=3.1.2
pylint>=2.11.1
python-dotenv>=0.19.0
fpdf2>=2.4.5
matplotlib>=3.4.3
seaborn>=0.11.2
httpx>=0.24.0
python-multipart>=0.0.5
slowapi>=0.1.9
